temp

+++ src/main/java/tsystems/javaschool/eCare/service/TariffServiceImpl.java
@@ -1,15 +1,21 @@
 package tsystems.javaschool.eCare.service;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import tsystems.javaschool.eCare.ECareException;
 import tsystems.javaschool.eCare.dao.TariffDAO;
 import tsystems.javaschool.eCare.model.Tariff;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Service
 public class TariffServiceImpl implements TariffService {
+
+    private static Logger logger = Logger.getLogger(TariffService.class);
 
     private TariffDAO tariffDAO;
 
@@ -20,31 +26,94 @@ public class TariffServiceImpl implements TariffService {
 
     @Override
     @Transactional
-    public List<Tariff> allTariffs() {
-        return tariffDAO.allTariffs();
+    public List<Tariff> getAllTariffs() {
+        logger.info("Get all tariffs from DB.");
+        List<Tariff> tariffs = tariffDAO.getAllTariffs();
+        //If DAO returns null method will throws an ECareException.
+        if (tariffs == null) {
+            ECareException ecx = new ECareException("Failed to get all tariffs from DB.");
+            logger.warn(ecx.getMessage(), ecx);
+            throw ecx;
+        }
+        logger.info("All tariffs obtained from DB.");
+        // Else method returns list of tariff entities.
+        return tariffs;
     }
 
     @Override
     @Transactional
-    public void add(Tariff tariff) {
-        tariffDAO.add(tariff);
+    public void add(Tariff tariff) throws ECareException {
+        logger.info("Add tariff " + tariff + " in DB.");
+        try {
+            tariffDAO.add(tariff);
+            logger.info("Tariff " + tariff + " added in DB.");
+        } catch (Exception ex) {
+            ECareException eCareException = new ECareException("Failed to add tariff " + tariff + " in DB.", ex);
+            logger.warn(eCareException.getMessage(), ex);
+            throw eCareException;
+        }
     }
 
     @Override
     @Transactional
-    public void edit(Tariff tariff) {
-        tariffDAO.edit(tariff);
+    public void edit(Tariff tariff) throws ECareException {
+        logger.info("Update tariff " + tariff + " in DB.");
+        try {
+            tariffDAO.edit(tariff);
+            logger.info("Tariff " + tariff + " updated in DB.");
+        } catch (Exception ex) {
+            ECareException eCareException = new ECareException("Failed to update tariff " + tariff + " in DB.", ex);
+            logger.warn(eCareException.getMessage(), ex);
+            throw eCareException;
+        }
     }
 
     @Override
     @Transactional
-    public void delete(Tariff tariff) {
+    public void delete(Long id) throws ECareException {
+        logger.info("Delete tariff with id: " + id + " from DB.");
+        Tariff tariff = tariffDAO.getTariffById(id);
+        //If DAO returns null method will throws an ECareException.
+        if (tariff == null) {
+            ECareException ecx = new ECareException("Tariff with id = " + id + " not exist.");
+            logger.warn(ecx.getMessage(), ecx);
+            throw ecx;
+        }
+        // Else tariff will be deleted from the database.
         tariffDAO.delete(tariff);
+        logger.info("Tariff " + tariff + " deleted from DB.");
     }
 
     @Override
     @Transactional
-    public Tariff getById(int id) {
-        return tariffDAO.getById(id);
+    public Tariff getTariffById(Long id) throws ECareException {
+        logger.info("Load tariff with id: " + id + " from DB.");
+        Tariff tariff = tariffDAO.getTariffById(id);
+        //If DAO returns null method will throws an ECareException.
+        if (tariff == null) {
+            ECareException ecx = new ECareException("Tariff with id = " + id + " not found.");
+            logger.warn(ecx.getMessage(), ecx);
+            throw ecx;
+        }
+        logger.info("Tariff " + tariff + " loaded from DB.");
+        //Else method returns tariff entity.
+        return tariff;
+    }
+
+    @Override
+    @Transactional
+    public void deleteAllTariffs() {
+        logger.info("Delete all tariffs from DB.");
+        tariffDAO.deleteAllTariffs();
+        logger.info("All tariffs deleted from DB.");
+    }
+
+    @Override
+    @Transactional
+    public Long getNumberOfTariffs() {
+        logger.info("Get number of tariffs in DB.");
+        Long number = tariffDAO.getSize();
+        logger.info(number + " of tariffs obtained fromDB.");
+        return number;
     }
 }
